pyeemd: allow running with larger num_imfs
If no more IMFs can be extracted but the residual is not monotonic,
several returned IMFs might be actually parts of the residual. However,
this is not really an error. Someone should devise a clear definition of
a residual, so libeemd can really stop when that condition has been
reached.